+++ monolib-dart/test/general/on_nullable_string_test.dart
@@ -7,6 +7,7 @@ void main() {
       expect(''.emptyToNull(), isNull);
       expect((null as String?).emptyToNull(), isNull);
       expect('hello'.emptyToNull(), 'hello');
+      expect('   '.emptyToNull(), '   ');
     });
   });
 }